* commit 'ad4d541fe2b8d8c9f2a04f19a01802371787a566': (82 commits)
  Be careful not to reset validity to true in subject validation
  Change validation message for duplicate slug (2)
  Give more immedidately-useful validation message for duplicate slug
  Limit length of the HTML fields to 10000 so as not to overflow data
  If JsonTree won't generate a result, remove any existing one
  If nothing to do on Sesame unimport, explicitly return success
  Use UTC for logstash log timestamps
  Specify MySQL database creation using utf8mb4
  Don't log info about file migration with error status
  Run concept transforms for PoolParty Project
  Migrate file and Sesame downloads using Registry download prefix
  During migration, don't add duplicate languages
  Be careful when migrating related entities not to create duplicates
  In JsonTree alert HTML, include version slug
  Migrate VIAFs with URLs that begin with https://
  If force workflow, do it, even if no other change to version metadata
  Rename workflow outcome's Result to ResultPair to avoid conflict
  Replace # with _ in upload filenames
  Support either integer or string "p" filter for search
  Add @SuppressWarnings("unused") for unused request and uriInfo parameters
  ...

+++ src/main/java/au/org/ands/vocabs/registry/model/AccessPointsModel.java
@@ -6,14 +6,17 @@ import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Set;
 
 import javax.persistence.EntityManager;
 import javax.ws.rs.core.MultivaluedHashMap;
 import javax.ws.rs.core.MultivaluedMap;
 
+import org.apache.commons.collections4.ListUtils;
 import org.apache.commons.collections4.sequence.CommandVisitor;
 import org.apache.commons.collections4.sequence.SequencesComparator;
 import org.apache.commons.lang3.BooleanUtils;
@@ -28,7 +31,9 @@ import au.org.ands.vocabs.registry.db.entity.AccessPoint;
 import au.org.ands.vocabs.registry.db.entity.ComparisonUtils;
 import au.org.ands.vocabs.registry.db.entity.Version;
 import au.org.ands.vocabs.registry.db.entity.Vocabulary;
+import au.org.ands.vocabs.registry.db.entity.clone.AccessPointClone;
 import au.org.ands.vocabs.registry.enums.AccessPointType;
+import au.org.ands.vocabs.registry.enums.ApSource;
 import au.org.ands.vocabs.registry.enums.SubtaskOperationType;
 import au.org.ands.vocabs.registry.enums.VocabularyStatus;
 import au.org.ands.vocabs.registry.model.sequence.AccessPointElement;
@@ -119,6 +124,19 @@ public class AccessPointsModel extends ModelBase {
                     AccessPointDAO.getDraftAccessPointListForVersion(em(),
                             versionId));
         }
+
+        // Now, take into account the fact that this may be invoked
+        // as part of a "refresh" done by VersionsModel.populateSubmodels().
+        // There may now (temporarily) be some draft AP rows for which
+        // the corresponding draft Version row has just been deleted.
+        Set<Integer> versionIds = new HashSet<>();
+        versionIds.addAll(currentVersions.keySet());
+        versionIds.removeAll(draftVersions.keySet());
+        for (Integer versionId : versionIds) {
+            draftAPs.addAll(versionId,
+                    AccessPointDAO.getDraftAccessPointListForVersion(em(),
+                            versionId));
+        }
     }
 
     /** {@inheritDoc} */
@@ -132,8 +150,8 @@ public class AccessPointsModel extends ModelBase {
         description.add("AP | Vocabulary; Id: " + vocabularyId());
         if (currentAPs != null) {
             for (Integer vId : currentAPs.keySet()) {
-                for (AccessPoint ap
-                        : currentAPs.get(vId)) {
+                for (AccessPoint ap : ListUtils.emptyIfNull(
+                        currentAPs.get(vId))) {
                     description.add("AP | Current version has AP; "
                             + "V Id, AP Id: " + vId + ","
                             + ap.getAccessPointId());
@@ -142,8 +160,8 @@ public class AccessPointsModel extends ModelBase {
         }
         if (draftAPs != null) {
             for (Integer vId : draftAPs.keySet()) {
-                for (AccessPoint ap
-                        : draftAPs.get(vId)) {
+                for (AccessPoint ap : ListUtils.emptyIfNull(
+                        draftAPs.get(vId))) {
                     description.add("AP | Draft version has AP; "
                             + "V Id, AP Id: " + vId + ","
                             + ap.getAccessPointId());
@@ -188,8 +206,12 @@ public class AccessPointsModel extends ModelBase {
             Integer versionId = outputVersion.getId();
             List<au.org.ands.vocabs.registry.schema.vocabulary201701.
             AccessPoint> outputAPList = outputVersion.getAccessPoint();
-            for (AccessPoint ap : currentAPs.get(versionId)) {
-                outputAPList.add(apDbMapper.sourceToTarget(ap));
+            List<AccessPoint> currentAPList = currentAPs.get(versionId);
+            // It shouldn't be empty, but just in case, do a null check.
+            if (currentAPList != null) {
+                for (AccessPoint ap : currentAPList) {
+                    outputAPList.add(apDbMapper.sourceToTarget(ap));
+                }
             }
         }
     }
@@ -221,8 +243,12 @@ public class AccessPointsModel extends ModelBase {
             Integer versionId = outputVersion.getId();
             List<au.org.ands.vocabs.registry.schema.vocabulary201701.
             AccessPoint> outputAPList = outputVersion.getAccessPoint();
-            for (AccessPoint ap : draftAPs.get(versionId)) {
-                outputAPList.add(apDbMapper.sourceToTarget(ap));
+            List<AccessPoint> draftAPList = draftAPs.get(versionId);
+            // It shouldn't be empty, but just in case, do a null check.
+            if (draftAPList != null) {
+                for (AccessPoint ap : draftAPList) {
+                    outputAPList.add(apDbMapper.sourceToTarget(ap));
+                }
             }
         }
     }
@@ -231,7 +257,7 @@ public class AccessPointsModel extends ModelBase {
     @Override
     protected void deleteOnlyCurrent() {
         for (Integer vId : currentAPs.keySet()) {
-            for (AccessPoint ap : currentAPs.get(vId)) {
+            for (AccessPoint ap : ListUtils.emptyIfNull(currentAPs.get(vId))) {
                 List<Subtask> subtaskList =
                         WorkflowMethods.deleteAccessPoint(ap);
                 if (subtaskList == null) {
@@ -359,7 +385,7 @@ public class AccessPointsModel extends ModelBase {
     @Override
     protected void deleteDraftDatabaseRows() {
         for (Integer vId : draftAPs.keySet()) {
-            for (AccessPoint ap : draftAPs.get(vId)) {
+            for (AccessPoint ap : ListUtils.emptyIfNull(draftAPs.get(vId))) {
                 AccessPointDAO.deleteAccessPoint(em(), ap);
             }
         }
@@ -397,13 +423,16 @@ public class AccessPointsModel extends ModelBase {
     /** {@inheritDoc} */
     @Override
     protected void notifyDeleteDraftVersion(final Integer versionId) {
-        for (AccessPoint ap : draftAPs.get(versionId)) {
-            // Just delete the row;
-            // for a draft instance, no workflow is applied.
-            AccessPointDAO.deleteAccessPoint(em(), ap);
+        List<AccessPoint> draftAPList = draftAPs.get(versionId);
+        if (draftAPList != null) {
+            for (AccessPoint ap : draftAPList) {
+                // Just delete the row;
+                // for a draft instance, no workflow is applied.
+                AccessPointDAO.deleteAccessPoint(em(), ap);
+            }
+            // Remove from our own records.
+            draftAPs.remove(versionId);
         }
-        // Remove from our own records.
-        draftAPs.remove(versionId);
     }
 
     /** {@inheritDoc}
@@ -427,7 +456,7 @@ public class AccessPointsModel extends ModelBase {
     }
 
     /** Make the database's draft view of the
-     * VocabularyRelatedEntities match updatedVocabulary.
+     * access points match updatedVocabulary.
      * @param updatedVocabulary The updated vocabulary.
      */
     private void applyChangesDraft(
@@ -438,9 +467,12 @@ public class AccessPointsModel extends ModelBase {
         List<AccessPointElement> existingDraftSequence = new ArrayList<>();
         for (Entry<Integer, List<AccessPoint>> apList : draftAPs.entrySet()) {
             Integer versionId = apList.getKey();
-            for (AccessPoint ap : apList.getValue()) {
-                existingDraftSequence.add(new AccessPointElement(
-                    versionId, ap.getAccessPointId(), ap, null));
+            for (AccessPoint ap : ListUtils.emptyIfNull(apList.getValue())) {
+                // Only consider those access points with source=USER.
+                if (ap.getSource() == ApSource.USER) {
+                    existingDraftSequence.add(new AccessPointElement(
+                            versionId, ap.getAccessPointId(), ap, null));
+                }
             }
         }
         Collections.sort(existingDraftSequence);
@@ -459,10 +491,14 @@ public class AccessPointsModel extends ModelBase {
                     Integer versionId = version.getId();
                     version.getAccessPoint().forEach(
                             ap -> {
-                                updatedSequence.add(new AccessPointElement(
-                                        versionId, ap.getId(), null, ap));
-                                if (ap.getId() != null) {
-                                    updatedAPs.put(ap.getId(), ap);
+                                // Only consider those access points with
+                                // source=USER.
+                                if (ap.getSource() == ApSource.USER) {
+                                    updatedSequence.add(new AccessPointElement(
+                                            versionId, ap.getId(), null, ap));
+                                    if (ap.getId() != null) {
+                                        updatedAPs.put(ap.getId(), ap);
+                                    }
                                 }
                             });
                 });
@@ -569,19 +605,27 @@ public class AccessPointsModel extends ModelBase {
                     }
                     // We don't need to create a new version Id, but
                     // we do need to add a draft row for it.
-                    Version draftVersion = draftVersions.get(versionId);
-                    versionsModel.workflowRequired(
-                            vocabularyModel.getDraftVocabulary(), draftVersion);
-                    TaskInfo taskInfo =
-                            versionsModel.getTaskInfoForVersion(versionId);
-                    Pair<AccessPoint, List<Subtask>> insertResult =
-                    WorkflowMethods.insertAccessPoint(em(), null, taskInfo,
-                            true, modifiedBy(), nowTime(), schemaAP);
-                    AccessPoint insertedAP = insertResult.getLeft();
-                    if (insertedAP != null) {
-                        draftAPs.add(versionId, insertedAP);
-                    }
-                    accumulateSubtasks(draftVersion, insertResult.getRight());
+                    AccessPoint draftAP = AccessPointClone.INSTANCE.clone(
+                            currentAP);
+                    draftAP.setModifiedBy(modifiedBy());
+                    TemporalUtils.makeDraft(draftAP);
+                    AccessPointDAO.saveAccessPoint(em(), draftAP);
+                    draftAPs.add(versionId, draftAP);
+                    // If we were doing workflow for drafts, might have
+                    // to do something like this:
+//                    Version draftVersion = draftVersions.get(versionId);
+//                    versionsModel.workflowRequired(
+//                        vocabularyModel.getDraftVocabulary(), draftVersion);
+//                    TaskInfo taskInfo =
+//                            versionsModel.getTaskInfoForVersion(versionId);
+//                    Pair<AccessPoint, List<Subtask>> insertResult =
+//                    WorkflowMethods.insertAccessPoint(em(), null, taskInfo,
+//                            true, modifiedBy(), nowTime(), schemaAP);
+//                    AccessPoint insertedAP = insertResult.getLeft();
+//                    if (insertedAP != null) {
+//                        draftAPs.add(versionId, insertedAP);
+//                    }
+//                    accumulateSubtasks(draftVersion, insertResult.getRight());
                 } else {
                     // Error: we don't know about this AP Id.
                     // (Well, it might be a historical version that
@@ -611,7 +655,7 @@ public class AccessPointsModel extends ModelBase {
     }
 
     /** Make the database's currently-valid view of the
-     * VocabularyRelatedEntities match updatedVocabulary.
+     * access points match updatedVocabulary.
      * @param updatedVocabulary The updated vocabulary.
      */
     private void applyChangesCurrent(
@@ -622,9 +666,12 @@ public class AccessPointsModel extends ModelBase {
         List<AccessPointElement> currentSequence = new ArrayList<>();
         for (Entry<Integer, List<AccessPoint>> apList : currentAPs.entrySet()) {
             Integer versionId = apList.getKey();
-            for (AccessPoint ap : apList.getValue()) {
-                currentSequence.add(new AccessPointElement(
-                    versionId, ap.getAccessPointId(), ap, null));
+            for (AccessPoint ap : ListUtils.emptyIfNull(apList.getValue())) {
+                // Only consider those access points with source=USER.
+                if (ap.getSource() == ApSource.USER) {
+                    currentSequence.add(new AccessPointElement(
+                            versionId, ap.getAccessPointId(), ap, null));
+                }
             }
         }
         Collections.sort(currentSequence);
@@ -651,10 +698,14 @@ public class AccessPointsModel extends ModelBase {
                     updatedVersions.put(versionId, version);
                     version.getAccessPoint().forEach(
                             ap -> {
-                                updatedSequence.add(new AccessPointElement(
-                                        versionId, ap.getId(), null, ap));
-                                if (ap.getId() != null) {
-                                    updatedAPs.put(ap.getId(), ap);
+                                // Only consider those access points with
+                                // source=USER.
+                                if (ap.getSource() == ApSource.USER) {
+                                    updatedSequence.add(new AccessPointElement(
+                                            versionId, ap.getId(), null, ap));
+                                    if (ap.getId() != null) {
+                                        updatedAPs.put(ap.getId(), ap);
+                                    }
                                 }
                             });
                 });
