+++ src/main/java/au/org/ands/vocabs/registry/model/VersionArtefactsModel.java
@@ -11,6 +11,7 @@ import javax.persistence.EntityManager;
 import javax.ws.rs.core.MultivaluedHashMap;
 import javax.ws.rs.core.MultivaluedMap;
 
+import org.apache.commons.collections4.ListUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -106,6 +107,23 @@ public class VersionArtefactsModel extends ModelBase {
                     VersionArtefactDAO.getDraftVersionArtefactListForVersion(
                             em(), versionId));
         }
+        /* The following is code copied from AccessPointsModel.populateModel()
+         * and adjusted for this class. We don't need it now, as we
+         * currently don't generate any VA database entities that are drafts.
+         * If this changes in future, revisit this code.
+        // Now, take into account the fact that this may be invoked
+        // as part of a "refresh" done by VersionsModel.populateSubmodels().
+        // There may now (temporarily) be some draft VA rows for which
+        // the corresponding draft Version row has just been deleted.
+        Set<Integer> versionIds = new HashSet<>();
+        versionIds.addAll(currentVersions.keySet());
+        versionIds.removeAll(draftVersions.keySet());
+        for (Integer versionId : versionIds) {
+            draftVAs.addAll(versionId,
+                    VersionArtefactDAO.getDraftVersionArtefactListForVersion(
+                            em(), versionId));
+        }
+        */
     }
 
     /** {@inheritDoc} */
@@ -119,8 +137,8 @@ public class VersionArtefactsModel extends ModelBase {
         description.add("VA | Vocabulary; Id: " + vocabularyId());
         if (currentVAs != null) {
             for (Integer vId : currentVAs.keySet()) {
-                for (VersionArtefact va
-                        : currentVAs.get(vId)) {
+                for (VersionArtefact va : ListUtils.emptyIfNull(
+                        currentVAs.get(vId))) {
                     description.add("VA | Current version has VA; "
                             + "V Id, VA Id: " + vId + ","
                             + va.getVersionArtefactId());
@@ -129,8 +147,8 @@ public class VersionArtefactsModel extends ModelBase {
         }
         if (draftVAs != null) {
             for (Integer vId : draftVAs.keySet()) {
-                for (VersionArtefact va
-                        : draftVAs.get(vId)) {
+                for (VersionArtefact va : ListUtils.emptyIfNull(
+                        draftVAs.get(vId))) {
                     description.add("VA | Draft version has VA; "
                             + "V Id, VA Id: " + vId + ","
                             + va.getVersionArtefactId());
@@ -176,7 +194,8 @@ public class VersionArtefactsModel extends ModelBase {
     @Override
     protected void deleteOnlyCurrent() {
         for (Integer vId : currentVAs.keySet()) {
-            for (VersionArtefact va : currentVAs.get(vId)) {
+            for (VersionArtefact va : ListUtils.emptyIfNull(
+                    currentVAs.get(vId))) {
                 List<Subtask> subtaskList =
                         WorkflowMethods.deleteVersionArtefact(va);
                 if (subtaskList == null) {
@@ -262,11 +281,13 @@ public class VersionArtefactsModel extends ModelBase {
     /** {@inheritDoc} */
     @Override
     protected void deleteOnlyDraft() {
-        // TO DO: since we really shouldn't have any draft VAs,
+        // Since we really shouldn't have any draft VAs,
         // once we have correctly applied workflow processing in this
         // class, there should be nothing more to do, so delete
         // the following row (and the body of the deleteDraftDatabaseRows()
         // method itself).
+        // Nevertheless, leave it for now, and we even now have
+        // some tests that confirm that such a cleanup happens!
         deleteDraftDatabaseRows();
     }
 
@@ -275,8 +296,11 @@ public class VersionArtefactsModel extends ModelBase {
     protected void deleteDraftDatabaseRows() {
         // Hmm, this isn't right. VAs are system-generated, so
         // we shouldn't have database deletion in this class.
+        // Nevertheless, leave it for now, and we even now have
+        // some tests that confirm that such a cleanup happens!
         for (Integer vId : draftVAs.keySet()) {
-            for (VersionArtefact va : draftVAs.get(vId)) {
+            for (VersionArtefact va : ListUtils.emptyIfNull(
+                    draftVAs.get(vId))) {
                 VersionArtefactDAO.deleteVersionArtefact(em(), va);
             }
         }
@@ -315,15 +339,20 @@ public class VersionArtefactsModel extends ModelBase {
     /** {@inheritDoc} */
     @Override
     protected void notifyDeleteDraftVersion(final Integer versionId) {
-        for (VersionArtefact ap : draftVAs.get(versionId)) {
-            // Just delete the row;
-            // for a draft instance, no workflow is applied.
-            // Hmm, this isn't right. VAs are system-generated, so
-            // we shouldn't have database deletion in this class.
-            VersionArtefactDAO.deleteVersionArtefact(em(), ap);
+        List<VersionArtefact> draftVAList = draftVAs.get(versionId);
+        if (draftVAList != null) {
+            for (VersionArtefact ap : draftVAList) {
+                // Just delete the row;
+                // for a draft instance, no workflow is applied.
+                // Hmm, this isn't right. VAs are system-generated, so
+                // we shouldn't have database deletion in this class.
+                // Nevertheless, leave it for now, and we even now have
+                // some tests that confirm that such a cleanup happens!
+                VersionArtefactDAO.deleteVersionArtefact(em(), ap);
+            }
+            // Remove from our own records.
+            draftVAs.remove(versionId);
         }
-        // Remove from our own records.
-        draftVAs.remove(versionId);
     }
 
     /** {@inheritDoc}
